requirements.txt
streamlit
altair
pandas
numpy
vega_datasets
rdkit
rdkit.Chem
sklearn.decomposition
rdkit.Chem.Draw